comment fixes I

+++ server/services/paymentPlan.service.js
@@ -1,6 +1,7 @@
 import PaymentPlan from '../models/paymentPlan.model';
 import { ErrorHandler } from '../helpers/errorHandler';
 import httpStatus from '../helpers/httpStatus';
+import { getPropertyById, updateProperty } from './property.service';
 
 export const getPaymentPlanById = async (id) => PaymentPlan.findById(id).select();
 
@@ -17,6 +18,12 @@ export const deletePaymentPlan = async (id) => {
   const plan = await getPaymentPlanById(id).catch((error) => {
     throw new ErrorHandler(httpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', error);
   });
+  if (plan.propertiesAssignedTo.length > 0) {
+    throw new ErrorHandler(
+      httpStatus.PRECONDITION_FAILED,
+      'Cannot delete payment plan with properties assigned',
+    );
+  }
   try {
     return PaymentPlan.findByIdAndDelete(plan.id);
   } catch (error) {
@@ -25,3 +32,45 @@ export const deletePaymentPlan = async (id) => {
 };
 
 export const getAllPaymentPlans = async () => PaymentPlan.find({});
+
+export const updatePaymentPlan = async (updatedPaymentPlan) => {
+  const plan = await getPaymentPlanById(updatedPaymentPlan.id).catch((error) => {
+    throw new ErrorHandler(httpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', error);
+  });
+  try {
+    return PaymentPlan.findByIdAndUpdate(plan.id, updatedPaymentPlan, { new: true });
+  } catch (error) {
+    throw new ErrorHandler(httpStatus.BAD_REQUEST, 'Error updating payment plan', error);
+  }
+};
+
+export const assignPaymentPlanToProperty = async ({ paymentPlanId, propertyId }) => {
+  const property = await getPropertyById(propertyId).catch((error) => {
+    throw new ErrorHandler(httpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', error);
+  });
+  const plan = await getPaymentPlanById(paymentPlanId).catch((error) => {
+    throw new ErrorHandler(httpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', error);
+  });
+
+  try {
+    const propertyToUpdate = {
+      id: property.id,
+      paymentPlan: plan.id,
+    };
+    const updatedProperty = await updateProperty(propertyToUpdate).catch((error) => {
+      throw new ErrorHandler(httpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', error);
+    });
+    if (updatedProperty) {
+      return PaymentPlan.findByIdAndUpdate(plan.id, {
+        $push: { propertiesAssignedTo: property.id },
+      });
+    }
+    return new ErrorHandler(httpStatus.BAD_REQUEST, 'Error assigning payment plan to property');
+  } catch (error) {
+    throw new ErrorHandler(
+      httpStatus.BAD_REQUEST,
+      'Error assigning payment plan to property',
+      error,
+    );
+  }
+};
